extract createStatementData to another file
