+++ memo-backend/test/service/statement.v4.test.js
@@ -1,0 +1,12 @@
+const invoice = require('../../service/statement/invoinces.json');
+const plays = require('../../service/statement/plays.json');
+const statement = require('../../service/statement/statement.v4');
+
+const result = 'Statement for BigCo\n '
+                + 'Hamlet: $650.00 (55 seats)\n '
+                + 'As You Like It: $580.00 (35 seats)\n '
+                + 'Othello: $500.00 (40 seats)\n'
+                + 'Amount owed is $1,730.00\nYou earned 47 credits\n';
+test('ok', () => {
+  expect(statement.statement(invoice, plays)).toBe(result);
+});
